Split ts-resolvers test into smaller chunks to make it easier to target
