+++ packages/plugins/typescript/resolvers/tests/ts-resolvers.federation.spec.ts
